Add files via upload
Lasso and Ridge, when run using glmnet, have their folds set at random. Because of this, the models change LITERALLY EVERY TIME, even if I set a seed. So that sucks. But the way around it is to train and minimize lambda a bunch of times to get the optimal lambda by running 100 times.